Remove Build Settings section from publish serverless wizard.
The Build Settings section was valid in the original version but since 2018 the framework is determined based on the Lambda runtime or project file. Since deploying through template can now involve multiple project with different framework the Build Settings doesn't make sense.

I also removed a metric that was captured because the framework is no longer being set. The metric was was also not guaranteed to be sent before again due the fact the framework set in the wizard is not the same as determine by the template. Also the template can now be outside of the project folder.