Fix compiling with msbuild
This change gets the repo to successfully compile the toolkit with msbuild when starting from scratch.

The compile failures are caused when toolkit compilation tries to reference nuget packages that target a different framework. This happens because of a combination of: HintPath references to the solutions\packages\ folder, and older packages.config style nuget package references.

The AWS Studio and Build Tasks related projects have these older package references, and are a part of the msbuild targets. I've updated the package reference style (maintaining the existing referenced versions), and ensured they still compile.

Ultimately, a pass is necessary to change all of the projects to stop using hint paths, and normalize some of the referenced packages, but that should be a separate dedicated task, as those changes may have an impact on the resulting VSIX output.

Changes to msbuild:

* removed the target steps that build for VS 2013 and VS 2015 since those are no longer supported
* build-vstoolkit is now the main "compile the toolkit" target once more, and it builds the VS 2017 toolkit. (build-vstoolkit-2017 can still be used)
* building the toolkit now performs the nuget restore step
* AWS Studio is no longer part of the main compilation paths. It has its own target, build-awsstudio
* removed some unused Targets

Other Notes:

* The AWS Studio solution compiles, but does not run. This is out of scope.

Test plan:

* Using Visual Studio, Clear out NuGet cache
* force clean the repo (fully removing all bin/obj/etc )
* msbuild buildtools\build.proj /t:build-vstoolkit
* See that the compile succeeds without errors