Fix for deploying to Beanstalk and getting a RED health check (#45)
When the toolkit creates a service role during a Beanstalk deployment, the role does not have elasticloadbalancing:DescribeTargetHealth permissions. This causes multi instance ALB health to get reported as RED.

This change generates a role that uses the managed policy arn:aws:iam::aws:policy/service-role/AWSElasticBeanstalkService. The following permissions are no longer in the role as a result:

- ec2:DescribeInstanceStatus
- ec2:GetConsoleOutput

Out of scope: Some sort of detection and/or notification for users that already have a role, which does not contain the elasticloadbalancing:DescribeTargetHealth permission.

Tested by...

- deleting all aws-elasticbeanstalk-* roles from my account
- deploying an asp.net core app to beanstalk, with a multi instance ALB
- app deployed, successfully showed a GREEN health status