Toolkit now uses the Win32 MessageBox to fix pop-under and similar issues when showing messages from threadpool running code
This change fixes an issue where message boxes may display one or more of these behaviors:

- they show up under other windows
- they are not truly modal, and allow focus to change within the IDE

This is more likely to happen when code running in threadpools calls to show a message, and the user is currently focused on a non-IDE window.

Unfortunately, the WPF MessageBox class requires a parent Window, and does not allow callers to provide an owner handle. The VS API only exposes owner handles. Using the Win32 method solves this problem.