Fix beanstalk deploy "security group does not exist" issue when using RDS security group
This change fixes an issue where trying to deploy .NET Core applications to Beanstalk (specifically Linux) and using an RDS security group would result in a "Security group does not exist" error. (Issue 107)

It appears there was some "LaunchIntoVPC" logic that needed to be accounted for earlier on in the deploy with .NET Core processing.