Added some encapsulation around PersistenceSettings (#90)
Newer versions of the AWSSDK limit what is accessible through the PersistenceSettings instance. These changes have been made to eliminate or reduce the compiler errors that would occur from updating the referenced AWSSDK versions.

This change increases the amount of PersistenceSettings encapsulation behind SettingsPersistence in case we decide to replace it in the future.

Performing an "instance of" check on the PersistenceSettings Instance is based on guidance from the sdk team.

The change was made by updating the referenced AWS SDK versions, fixing all compilation errors, then reverting the SDK version change.