ECS Deploy metrics (#104)
This change migrates ECR/ECS deployments to use the new Telemetry system:
* Deploy image to ECR
* Deploy ECS Tasks
* Deploy ECS Scheduled Tasks
* Deploy ECS Services

The linecount is on the higher side for this PR. The reason was that I added test coverage to verify that success/fail metrics would be emitted. The existing xxxWorker classes do not contain an encapsulation against docker commands, so I introduced an IEcsDeploy interface to wrap existing implementations. This lets me test the metrics portions of the worker classes without performing any of the actual deploy operations. Much of this code got moved around within their respective files, which runs up the change counts.