Remove ifdef conditionals relating to VS versions (#195)
At this time, the repo builds one Toolkit VSIX that supports VS2017 and VS2019. This change removes all ifdef code that relates to VS versioning. Much of this was necessary when the repo used to produce different Toolkits for separate versions of VS like 2015, 2013, and pre-2013.

This change is being made as part of a series of changes that will remove unnecessary version branching logic, with the ultimate goal of getting the repo compiling within VS 2019.