New RegionProvider (not integrated) (#225)
This change adds a RegionProvider, responsible for resolving all partition-region-service information for the toolkit. It uses the new endpoints.json file, and is intended to phase out ServiceEndPoints.xml.

This change contains the RegionProvider implementation, but it is not integrated with the Toolkit. An upcoming set of changes will integrate this provider with the toolkit and remove usage of the xml file.

Also:

Moved TestUtil to TestResources, made slight adjustments to promote reuse.