Fix CloudFormation errors with short form yaml; Make stack template more legible (#279)
When the Toolkit loads yaml based template files, the yaml parser raises an exception when it encounters short form notation like !Ref. This meant that users could not view stacks that were created with short-form yaml, nor could they publish yaml CloudFormation templates containing short form.

This change adjusts the yaml parsing to only load the necessary fields in a more generalized key-value manner, avoiding a file-wide parse (and avoiding the exception).

A bonus change was made to make the "Template" tab of the stack viewer a little more legible. I maintained a monospace font, but I made it bigger, and non-bold.