Add AccountManager as a fallback to retrieve account number (#322)
With the removal of the Account Number field from Credential UI's, the account number field associated with AccountViewModels can be blank/null, which leads to incorrect arns being copied for resources on drag and drop. This change adds AccountManager as a fallback to retrieve account Id for a given profile by making a get:StsCallerIdentity call.
Note: This change has been made as a temporary stop-gap till the new credential system is integrated. The feature branch already contains a way to handle this scenario.