Remove unused third party dlls (#349)
This change removes some unused dlls from the thirdparty folder. There is no net change to the VSIX contents.

I tried to remove WPFToolkit but hit an exception at runtime when looking at Beanstalk graphs, so that one had to stay in. I added a comment to warn about this.