Get Solution Explorer's current selection (#362)
The Toolkit contains some code in VSUtility.cs that queries Visual Studio for the currently selected item in the Solution Explorer. This change makes that code testable with mocks and brings that code under test. This change will be used in some upcoming work that involves getting the selected items, and I want that upcoming work to have test coverage.

This change creates a new test project - AWSToolkit.VisualStudio.Shared.Tests - which contains tests for the AWSToolkit.VisualStudio.Shared module. The reason these tests aren't added into other projects is to limit the spread of references to the VS SDK from other code. Unfortunately, VS SDK references were added to AWSToolkit.Tests.Common to support the addition of some test fixtures that will be used both in the above test project, and in a new publish related test project (being worked on in a feature branch).