List vsix contents on compilation (#403)
This change introduces a new build target list-vsix-contents. This target lists the contents of the toolkit VSIX file, and writes the list to a text file (_vsixContents.txt).

This operation has been added to AWSToolkitPackage.csproj. When you compile this project, it will run the list contents step after producing the VSIX file.

The motivation for this change is to help inspect/diagnose the extension more easily when adding new projects or dependencies to the Toolkit. It will also be handy as we restructure the projects to accommodate Visual Studio 2022.