"AWS Toolkit" menu, with commands to View User Guide and Create Issue (#415)
This change sets up an "AWS Toolkit" menu, a place to centralize all top level actions for the extension. The menu is placed within the "Extensions" menu, however this menu was only added in VS 2019. For VS 2017, the "AWS Toolkit" menu is added to the "Tools" menu. As a result, we have to define duplicate sets of menu item groups and commands. The VS 2017 related groups and commands are suffixed with `2017`, as those would get dropped in the future, if/when support for VS 2017 is removed.

The new menu has been seeded with two commands: one to create an issue on GitHub, and another to view the User Guide. Over time, this menu will fill in with additional commands (for example, view logs, send feedback).

Additional changes: small typo and spacing fixes were also made in the file