Writes Metrics to OutputWindow when ShowMetricsOutputWindow is set to true (#444)
This change adds an output panel, where the Toolkit writes all of the metrics it is logging. This helps Toolkit development by providing a way to visually see what metrics are intended to be sent to the backend. This is a faster loop than waiting for the metrics to get processed by the backend, but is not intended to be a replacement to backend verification.

To enable this feature, set `ShowMetricsOutputWindow` to true in the `MistSettings.json` file, prior to starting the Toolkit. When the Toolkit is started, there will be an OutputWindow dedicated to showing metrics details. This field defaults to false, in which case there will be no metrics OutputWindow.