Remove unused "Show me How" content from First Run dialog (#478)
When the "First Run" dialog was created, some content was placed in the view to show you how to view the AWS Explorer, and how to deploy to Lambda/Beanstalk. Except the content was marked hidden (collapsed). The content was never changed to be set visible, and there is no code that references it. The content appears to have included light and dark theme images, however only the light versions were ever referenced.

This change removes all of this never before seen footage. There's a good idea here, but the visuals will become dated.