Move ReleaseNotes.txt (#485)
This change moves ReleaseNotes.txt to a new folder, and updates the related changelog scripting to work with that new location.

The location "vspackages/AwsToolkitPackageResources" is a new location where resources from the AWSToolkitPackage project will be relocated in a separate change, in preparation to support VS 2022. This change was performed in advance because it had scripting associated with the file relocation.

This change was tested by generating release notes, producing a Release vsix, launching the vsix installer, and confirming that the "Release Notes" link shows the produced contents. If this link did not show the expected contents, the file would not have been successfully included in the extension.