Move AwsToolkit.VsSdk.Common code into a Shared Project (#502)
This change introduces a Shared Project AwsToolkit.VsSdk.Common.Shared. The code from AwsToolkit.VsSdk.Common has been relocated into the shared project.

Shared Projects are like symbolic links. If Shared Project S contains file X, and Project A has a reference to shared project S, Project A acts as if it contains file X.

The motivation for this change is to prepare to support VS 2022 extensions. Supporting VS 2022 requires separate projects for VS 2022 and for VS 2019 (and earlier) for any project that accesses the VS SDK.

References: 
* Shared Projects: https://docs.microsoft.com/en-us/xamarin/cross-platform/app-fundamentals/shared-projects?tabs=windows
* Migrating extensions to support VS 2022: https://docs.microsoft.com/en-us/visualstudio/extensibility/migration/update-visual-studio-extension?view=vs-2022#use-shared-projects-for-multi-targeting