Move AwsToolkitPackage code into a Shared Project (#504)
This change introduces a Shared Project AwsToolkitPackage.Shared. The code from AwsToolkitPackage has been relocated into the shared project.

Shared Projects are like symbolic links. If Shared Project S contains file X, and Project A has a reference to shared project S, Project A acts as if it contains file X.

The motivation for this change is to prepare to support VS 2022 extensions. Supporting VS 2022 requires separate projects for VS 2022 and for VS 2019 (and earlier) for any project that accesses the VS SDK.

References: 
* Shared Projects: https://docs.microsoft.com/en-us/xamarin/cross-platform/app-fundamentals/shared-projects?tabs=windows
* Migrating extensions to support VS 2022: https://docs.microsoft.com/en-us/visualstudio/extensibility/migration/update-visual-studio-extension?view=vs-2022#use-shared-projects-for-multi-targeting


This change also fixes an issue where CloudFormation projects and CloudFormation Tempates were not showing the CloudFormation icon in the Solution Explorer, caused by loading an incorrect path.

Change Verifications:
* ensure the contents of the generated VSIX file did not add new files or remove existing files
* run existing test suite
* Ran the Toolkit and exercised commands, deployments, AWS Explorer without issue
  * Checked that embedded resources like images are still showing in the Toolkit