Support for Visual Studio theming (#512)
This change introduces an abstraction that allows us to apply Visual Studio themes to our UIs. By applying a dependency property (ToolkitThemes.UseToolkitTheme="True") a UIs controls will gain the Visual Studio styling, so that their sizing, color, and general look and feel matches Visual Studio's currently loaded theme.

The UIs are currently set up with extremely explicit styling, and much of that styling was hardcoded against visual appearances from an older version of Visual Studio (2012?). 

UIs can use this new abstraction by:
* removing existing styles from a UI
* remove the code that loads the old toolkit theming (ResourceDictionary Themes/Generic.xaml)
* Adding the following lines to the window/control:
```
    xmlns:themes="clr-namespace:Amazon.AWSToolkit.Themes"
    themes:ToolkitThemes.UseToolkitTheme="True"
```

When we want to use a specific color intent, we need to create an appropriate key in `IToolkitThemeBrushKeys`, and associate it with a specific Brush, generally from the Visual Studio Shared Colors (https://docs.microsoft.com/en-us/visualstudio/extensibility/ux-guidelines/shared-colors-for-visual-studio?view=vs-2019). `HintText` was added as an initial example, and more keys will be added as needed.

This change is accomplished using Community toolkit for Visual Studio extensions (Community.VisualStudio.Toolkit - https://github.com/VsixCommunity/Community.VisualStudio.Toolkit). What is Community.VisualStudio.Toolkit? It is a project led by a Visual Studio product manager, where a community of extension authors contribute towards abstracting out portions of the Visual Studio SDK that are cumbersome, outdated, or undifferentiated. We can elect to leverage more of these features going forward, as long as we maintain a proper isolation between most of the toolkit and the VSSDK/CommunityToolkit (eg: AWSToolkitShellProviderService is still where we would want to encapsulate things).

An abstraction is placed around the Community.VisualStudio.Toolkit theme handling to give us the chance to override any theming behavior, or to deliberately add custom styles for specific controls that are not handled by the community toolkit.

This change was verified by altering an existing view to use this new dependency property. A follow-up PR will be made that showcases this dependency property's usage. This change focuses strictly on the new abstraction, and is otherwise dormant.

The generated VSIX was examined, to ensure that the community toolkit dll and license file are the only additional files that appear in the VSIX.

Unfortunately, a test couldn't be added around ToolkitThemeProvider.cs, because the dependency property requires Visual Studio to be running in order to access the VS themes that are applied.