Apply Visual Studio Theming to Credentials dialog (#516)
This change uses the credentials dialog as a reference in how to apply the new theming introduced in #512.

The dialog has been adjusted to use the Visual Studio styles, and to match the Visual Studio layout guidelines for spacing. It illustrates how we won't need to apply styles to every control in a view. The dialog now automatically supports the user's selected Visual Studio theme as well.

Unfortunately, some additional changes were required to support these new themed views. The way the Toolkit generally shows modal dialogs is to have a more generic "host" window (eg: OkCancelDialogHost), where a custom UserControl is (dynamically) placed. The generic host window contains the buttons responsible for dismissing the window (eg: Ok, Cancel), and is abstracted from the rest of the Toolkit through IAWSToolkitShellProvider::ShowModal.

These host windows are set up to use the legacy (hardcoded) themes. Changing the host window to use new theming, but leaving the existing UserControls with legacy theming (or vice versa) results in a window that has an off-colored border around the control (and an off colored footer, where the buttons are located). So, identical support code was created, so that newly themed UserControls can be displayed properly. IAWSToolkitShellProvider::ShowInModalDialogWindow has been set up, which utilizes OkCancelDialogWindowHost. OkCancelDialogWindowHost is identical to OkCancelDialogHost, except that some of the code was cleaned up a little. OkCancelDialogWindowHost also makes use of the VS SDK based DialogWindow, which a is more recommended Window to use in extensions:

> Someone implementing a WPF dialog should derive from this class (to have consistent styling with other VS dialogs, and help support) To display the dialog invoke the ShowModal() function, and that will take care of correctly parenting the dialog in the shell, enabling modal state for the shell while the dialog is displayed, etc.

Reference: https://docs.microsoft.com/en-us/dotnet/api/microsoft.visualstudio.platformui.dialogwindow