Remove Lambda Handler component fields (#543)
This change removes the UI fields that contribute to the Lambda handler. 

These changes are being made:
* to make space in the UI for upcoming changes (which will add more controls)
* to simplify the dialog (through less fields)

The Lambda handler hint has been re-added to the UI, has been slightly reworded, and is selectable. Some of the Tooltip/Hint text was moved to the viewmodel and brought under test. 

My IDE appears to have autoformatted portions of the UploadFunctionViewModel file. That was not intentional, but was bound to happen, so I left it.