Add VS 2022 to version detection; move VsMinimumSupportedVersion (#647)
This change updates the code to identify if the running instance of Visual Studio is version 2022.

`VsMinimumSupportedVersion` was moved out of `AWSToolkit.Util` into a project that has different versions per VS Version, allowing us to conditionally declare what the minimum version is for an extension (eg: VS 2017+2019 Toolkit has a min ver of 2017, VS 2022 future Toolkit has a min ver of 2022). This is low impact, since the property is only used in one place.