Move common project directives from AWSToolkitPackage into sharable files (#654)
We will spin up a separate AWSToolkitPackage project to target VS 2022, and it will have many identical references to AWSToolkitPackage.csproj. We could do this by copying a lot of the csproj contents into a new project file, but this is challenging to keep in sync as both projects move forward.

This change takes project references that will be common to both projects, and moves them into separate files. These common files are then imported into AWSToolkitPackage.csproj, which is functionally equivalent, and gives us something to leverage in the (future) 2022 based project.

This change has added benefits:
* similar references have been bucketed together for context
* the line count of AWSToolkitPackage.csproj has been greatly reduced

This change was verified by:
* checking that the contents of a VSIX were the same before and after this change (this change introduces no new files and does not remove files)
* running the Toolkit and giving random features a check to ensure they work (making a lambda project, publishing a lambda, ...)