Start of VS 2022 Toolkit (#660)
This change introduces a new project (`AWSToolkitPackage.v17`) that represents the VS 2022 version of the AWS Toolkit. The project currently compiles, and hosts most (but not all) of the Toolkit's capabilities. Upcoming tasks will identify and fix any runtime-based incompatibilities that arise with the new 64-bit architecture of VS 2022.

The produced extension has the filename `AWSToolkitPackage.v17.vsix`

Some of the components not available in the VS 2022 Toolkit with this change:
* CloudFormation EditorExtensions - CloudFormation autocompletion
* All of the New Project and New Item integrations
* CodeCommit integration with TeamExplorer

The 2022 toolkit can currently be compiled from VS 2022 or VS 2019, but in the future, it will likely become a requirement to use VS 2022 to produce the VS 2022 Toolkit (when it uses more specific packages like TeamExplorer). The 2017/2019 toolkit can be compiled from VS 2017/2019/2022 at this time.

Here are the changes that support creating this initial project
* An extension GUID has been designated for VS 2022 Toolkit, and configured where necessary. The VS 2017/2019 Toolkit continues to use its own unique guid
* Same for the Extension's display name
* A build task `list-vsix-contents-2022` has been created (identical to `list-vsix-contents`) which will output the contents of the VS 2022 vsix file, to help while developing the extension
* The VS 2022 SDKs have some compatibility breaking changes to account for the shift to 64 bits. Adjustments have been made to a few code files, so that they compile for the 32 bit VS 2017/2019 Toolkit, and the 64 bit VS 2022 Toolkit. Places that we have to revisit after getting more of the 2022 Toolkit propped up have corresponding backlog items for tracking.
  * Reference: https://docs.microsoft.com/en-us/visualstudio/extensibility/migration/update-visual-studio-extension?view=vs-2022