Split project template references (#690)
Project Templates that have custom handlers leverage the VS SDK, which will require a split to support newer 64 bit VS 2022. Because each Toolkit extension will require VS-version specific project templates, this change adjusts the references from the AWSToolkitPackage projects to the template projects so that they reference vs-version specific template projects. The references in the VS 2022 (v17) file are commented out as placeholders to be brought online in upcoming changes.

`ProjectTemplates.v15.xml` is deliberately named 15 instead of 16 because it supports VS 2017 as the minimum version.

The motivation for this change is to prepare to support VS 2022 extensions. Supporting VS 2022 requires separate projects for VS 2022 and for VS 2019 (and earlier) for any project that accesses the VS SDK.

References: 
* Shared Projects: https://docs.microsoft.com/en-us/xamarin/cross-platform/app-fundamentals/shared-projects?tabs=windows
* Migrating extensions to support VS 2022: https://docs.microsoft.com/en-us/visualstudio/extensibility/migration/update-visual-studio-extension?view=vs-2022#use-shared-projects-for-multi-targeting

This change was tested by:
* Loading into VS 2019, verifying all AWS templates appear, and that I could create a C# Lambda project
* Loading into VS 2022, verifying no AWS templates appear (none are set up at this time)