Set up Lambda TemplateWizard project for 2022 (#691)
This change creates the VS 2022 version of the project that will be referenced by the (future) VS 2022 specific Lambda Project templates. The new project leverages the shared project created in #666 

`Microsoft.VSSDK.BuildTools` was also bumped to a more recent preview version.

The motivation for this change is to prepare to support VS 2022 extensions. Supporting VS 2022 requires separate projects for VS 2022 and for VS 2019 (and earlier) for any project that accesses the VS SDK.

References: 
* Shared Projects: https://docs.microsoft.com/en-us/xamarin/cross-platform/app-fundamentals/shared-projects?tabs=windows
* Migrating extensions to support VS 2022: https://docs.microsoft.com/en-us/visualstudio/extensibility/migration/update-visual-studio-extension?view=vs-2022#use-shared-projects-for-multi-targeting

This change was tested by:
* Compiling for VS 2019, and verifying the VSIX contents were unaffected
* Compiling for VS 2022, and verifying the VSIX contents gained this new `AWSToolkit.Lambda.TemplateWizards.v17.dll`