Generate test builds (#700)
This change makes it easier to generate (or obtain) VSIX builds that can be used for local or internal testing.

Previously, I would look up the current Toolkit version, come up with a semi-incremented value from that, try to remember the msbuild incantation to run the `update-version` task, then compile with the Release configuration. This is now all automated with the build task `build-test-toolkit`.

The intention is to run `build-test-toolkit` locally from a commit or branch of interest, whenever someone needs to produce and share a copy of the Toolkit for testing. This has also been hooked into GitHub Actions so that we can grab a compiled VSIX directly from PRs and commits.

These test builds are given a version that starts with the most recently released Toolkit version, then adds in the current month+day. This gives us Good Enough granularity to access test builds on a daily basis without having to worry about uninstalling an existing build before installing these auto-generated builds.

When the `build-test-toolkit` task is used locally, you can also provide an explicit version by using the optional `Version` parameter: `msbuild buildtools\build.proj /t:build-test-toolkit /p:Version=1.2.3.4`