Produce SHA 384 hashes with test builds (#750)
Whenever we share a test build publicly (for example, posting a build on a GitHub issue for customers to try out), it is a best practice to indicate a hash so that users can verify that someone from the team is providing a build. This change produces SHA 384 hashes with each test build. This makes it easier for us to share the hash values alongside any builds.