Remove Publish to AWS opt-in banners from VS 2022 (#761)
This change prevents the opt in banner for the new publish experience from showing up in VS 2022. The new experience is not set up to support 2022 yet, so we don't want users to see this.

This change was tested in VS 2022 to confirm the banners do not appear, and that they still appear in 2019.