Point VS 2022 Toolkit at VS 2022 Blueprints instead of VS 2019, now that they are published (#791)
This change sets up the VS 2022 Toolkit to use its own set of hostedfiles. Until now, it was using the .NET 5 based VS 2019 hostedfiles, because we hadn't published a set of 2022 hostedfiles. This means you can now access .NET 6 related blueprints from 2022.

This change was verified by trying to create a .NET 6 based custom lambda, and publishing it to AL2.