Telemetry Fixture, pass ToolkitContext to UploadGenericWorker (#818)
This change moves TelemetryFixture from Lambda specific tests over to Tests.Common, and makes it a part of ToolkitContextFixture. This way, other tests can leverage shared telemetry fixture functionality. The Lambda specific logic was retained in LambdaAssert.

This change is anchored to updating the UploadGenericWorker to accept ToolkitContext instead of a Telemetry Logger, which is setting up a future change to UploadGenericWorker, which will require ToolkitContext.