Fix error updating NodeJs Lambda functions when Lambda State is not ready (#819)
Recently the Lambda Service made a change that requires Create/Update function calls to occur only when the Lambda function is in a good State. This affected the Toolkit's ability to update NodeJs Lambda functions. This change now waits for the state to change, so that NodeJs function updates do not fail.

I had to add a ExecuteOnUIThread overload that will await an async task. I believe this is necessary because the code is already running from within another ExecuteOnUIThread scope. Otherwise the "wait for state" check does not actually wait.

Bug: 
* https://github.com/aws/aws-toolkit-visual-studio/issues/197

Reference: 
* https://aws.amazon.com/blogs/compute/tracking-the-state-of-lambda-functions/
* https://aws.amazon.com/blogs/compute/coming-soon-expansion-of-aws-lambda-states-to-all-functions/