Add IAM Role selection dialog (#825)
This change builds on #823 and adds an IAM Role selection dialog to the Toolkit. This dialog is not invoked from anywhere in the Toolkit, that is coming in a later change.

A new Toolkit plugin service `IAwsIamResources` has been created, which keeps the AWS SDK related logic out of places like VSSDK.Common, where the dialog resides.