Fix exception showing Telemetry banner when Toolkit is not fully initialized (#830)
When the Toolkit starts up, it tries to show an informational banner regarding usage data. The banner has to wait for the IDE to be available, and for the Toolkit to be initialized. There are situations where the Toolkit isn't initialized, and the banner startup code silently throws an exception, resulting in no banner being shown (at times when it should have been). One common way that this happens is when, in the previous VS session, the AWS Explorer is fully closed before closing down Visual Studio.

This change fixes that problem.