Add VPC Selection dialog (#833)
This change adds a VPC selection dialog to the Toolkit. This dialog is not invoked from anywhere in the Toolkit, that is coming in a later change.

A new Toolkit plugin service `IVpcRepository` has been created, which keeps the AWS SDK related logic out of places like VSSDK.Common, where the dialog resides.