Apply ConfigurationDetail validations to child details (#835)
This change takes ConfigurationDetail validations and applies them to child details when applicable. Previously the validation would only be applied to the detail that signaled a change, resulting in validations not getting surfaced when they are attributed to a child of that detail. Upcoming work relating to custom settings editors will explore treating a detail and its child details as one unit, and this change is necessary to surface validations back to the UI.

To make this change, `DeployToolController` was enhanced so that `ApplyConfigSettingsAsync` returns an abstraction (`ConfigurationDetailValidation`) around the validation results, instead of a raw dictionary. `ConfigurationDetailValidation` then bubbles up the call stacks, and is processed into the detail + child details in `PublishToAwsDocumentControl` - `SetTargetConfiguration`.

Also, some methods that I was working in were renamed to have an `Async` suffix.

This change was verified through:
* running the tests (including the publish integration test)
* performing deployments and redeployments from the Toolkit