Add ConfigurationDetail event to signal when self/child values change (#839)
This change gives `ConfigurationDetail` an event (`DetailChanged`) that is raised when it or its children's values change. The implementation has been set up so that derived classes can suppress events when desired (via `SuspendDetailChangeEvents`|`ResumeDetailChangeEvents`).

This change is being made to support an upcoming change where we want to be able to make a series of changes to child values "atomically" before signaling that the `ConfigurationDetail` object has changed.