Prevent entry to new publish experience without valid credentials (#845)
If you enter the new publish experience without valid credentials, the Publish experience shows an empty screen that you cannot interact with. Until that experience gains the ability to adjust its own credentials, this change shows an error message instead of launching the publish experience, if credentials are not valid. The message informs the user that they need to select valid credentials from the AWS Explorer prior to selecting "Publish to AWS".