Add EC2 Instance Type Selection Dialog (#859)
This change adds an EC2 InstanceType selection dialog to the Toolkit. It builds on #843, which loads the instance type data from AWS. This dialog is not invoked from anywhere in the Toolkit, that is coming in a later change.

Dialog features:
* Filter searches the Instance type and the architecture fields
* Using spaces in the filter will perform an AND search of all words. (eg: "arm micro" gives all instance types with "arm" and "micro" within their type and architecture fields
* The types shown in the dialog can be restricted to certain architecture(s) using the `Architectures` property. This allows us to use the dialog in workflows where a user should only select from a list of (for example) ARM only types.

This change is similar to the VPC selection dialog, added in #833.