Key-Value UserControls (#877)
This change introduces ViewModels and UserControls for working with a set of Key Value pairs. The sets are validated to ensure the Keys are unique. This change does not add these controls anywhere - that will come in later changes.

There is a read-only control, that shows the current collection of key-values. There are two controls let users edit the collection. One is form-based, and the other lets users work in a multiline text field to "batch" operate on the collection.