Change ConfigurationDetail.Type to use a fixed enumeration (#879)
`ConfigurationDetail.Type` was initially implemented as type `Type`. We have only needed this to serve as an enumeration so far, and haven't needed to leverage the actual type. This change updates the property to use an enum type, in advance of an upcoming change that introduces a new KeyValue type.