Add Key-Value settings editor to Publish to AWS experience (#883)
This change adds the ability to edit Publish to AWS settings defined as key-value collections, such as Beanstalk environment variables. It pulls together the controls and dialogs that were created in #877 and #881.

This change was done in preparation for a new deploy tool CLI version, which contains "KeyValue" types in recipes. Until then, the change will be visible in the UI, because there are no deploy targets containing settings with this type.