Prevent CodeCommit projects from rebuilding due to always copying unused image (#884)
This change reduces the amount of time the team explorer projects need to be re-built when building the solution. The cause is an image set to "copy always" (https://michaelscodingspot.com/visual-studio-keeps-rebuilding-projects-no-good-reason/). It turns out these images are no longer used due to a recent pass to update and centralize the repo's icons, so they have been removed.