Update Lambda tools to have .NET 6 mapping to Lambda .NET 6 runtime enum (#909)
Updates Lambda tools to have the .NET 6 framework mapping to the .NET 6 Lambda runtime enum. Confirmed with Lambda that `dotnet6` will be the enum value.

There was also a recent change to switch from our custom `NETCORE` preprocessor to the standard `NETCOREAPP3_1_OR_GREATER` preprocessor in the tools repository. This will have no affect on the toolkit because neither the old `NETCORE` or the standard `NETCOREAPP3_1_OR_GREATER` preprocessor are define in the toolkit project files.