All publish set-up metrics are passive except for the "all" stage (#922)
The `publish_setup` metric instruments stages of deploy CLI installation, verification, and showing the dialog. The installation and verification steps take place on Toolkit startup, and dialogs are shown when the user requests them. This metric is being changed to be passive, except for the final stage (setupStage All), which reflects how long the user waited from menu click to seeing the dialog.