Add build property NuGetDisableParallelProcessing (#923)
This change adds a build property `NuGetDisableParallelProcessing` that can be used with the `restore` build target. When used, it instructs NuGet to not restore packages in parallel.

This flag is not used by default. It is being added to see if this will improve build stability from CodeBuild. For now, verbose output is used with this mode, to help diagnose issues if they arise.