Refactor the publish target models (#924)
This change adjusts `PublishRecommendation` and `RepublishTarget` so that they have a common base class `PublishDestinationBase`. The intent is to avoid using different properties to represent when the user wants to deploy to a new or existing target.

The ViewModel will make an attempt to "remember" the last target you selected whenever you toggle between new/existing target views.